Reformatting files
